genai/requirements.txt
fastapi
uvicorn[standard]
pydantic
weaviate-client
aio-pika
uuid